+++ src/pages/directory/edit.rs
@@ -6,11 +6,13 @@
 
 use std::{sync::Arc, vec};
 
-use ahash::AHashMap;
+use ahash::{AHashMap, AHashSet};
+use base64::{engine::general_purpose, Engine};
 use humansize::{format_size, DECIMAL};
 use leptos::*;
 use leptos_router::{use_navigate, use_params_map};
 use pwhash::sha512_crypt;
+use rand::{distributions::Alphanumeric, thread_rng, Rng};
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -18,7 +20,7 @@ use crate::{
         form::{
             button::Button,
             input::{InputPassword, InputSize, InputText},
-            select::{CheckboxGroup, Select},
+            select::Select,
             stacked_badge::StackedBadge,
             stacked_input::StackedInput,
             tab::Tab,
@@ -78,6 +80,8 @@ pub fn PrincipalEdit() -> impl IntoView {
             "tenants" => PrincipalType::Tenant,
             "domains" => PrincipalType::Domain,
             "roles" => PrincipalType::Role,
+            "api-keys" => PrincipalType::ApiKey,
+            "oauth-clients" => PrincipalType::OauthClient,
             _ => PrincipalType::Individual,
         }
     });
@@ -87,12 +91,6 @@ pub fn PrincipalEdit() -> impl IntoView {
         .permissions()
         .has_access(Permission::TenantList);
     let principals: RwSignal<Arc<PrincipalMap>> = create_rw_signal(Arc::new(AHashMap::new()));
-    let permissions = create_memo(move |_| {
-        PERMISSIONS
-            .iter()
-            .map(|(id, name)| (id.to_string(), name.to_string()))
-            .collect::<Vec<_>>()
-    });
 
     let fetch_principal = create_resource(
         move || params.get().get("id").cloned().unwrap_or_default(),
@@ -115,14 +113,17 @@ pub fn PrincipalEdit() -> impl IntoView {
                     PrincipalType::List,
                 ][..],
                 PrincipalType::Domain => &[PrincipalType::Tenant][..],
-                PrincipalType::Tenant => &[PrincipalType::Role][..],
+                PrincipalType::Tenant | PrincipalType::ApiKey => &[PrincipalType::Role][..],
                 PrincipalType::Role => &[PrincipalType::Role, PrincipalType::Tenant][..],
                 PrincipalType::List => &[
                     PrincipalType::Individual,
                     PrincipalType::Group,
                     PrincipalType::Tenant,
                 ][..],
-                PrincipalType::Resource | PrincipalType::Location | PrincipalType::Other => &[][..],
+                PrincipalType::Resource
+                | PrincipalType::Location
+                | PrincipalType::Other
+                | PrincipalType::OauthClient => &[][..],
             };
             let mut fetch_types = String::new();
             for typ in needed_types {
@@ -162,6 +163,21 @@ pub fn PrincipalEdit() -> impl IntoView {
                                 "tenant-admin".to_string(),
                                 "user".to_string(),
                             ]);
+                        }
+                        PrincipalType::Group => {
+                            principal.enabled_permissions = PrincipalValue::StringList(vec![
+                                "email-send".to_string(),
+                                "email-receive".to_string(),
+                            ]);
+                        }
+                        PrincipalType::ApiKey => {
+                            principal.secrets = PrincipalValue::StringList(vec![thread_rng()
+                                .sample_iter(Alphanumeric)
+                                .take(30)
+                                .map(char::from)
+                                .collect::<String>()]);
+                            principal.enabled_permissions =
+                                PrincipalValue::StringList(vec!["authenticate".to_string()]);
                         }
                         _ => {}
                     }
@@ -310,6 +326,12 @@ pub fn PrincipalEdit() -> impl IntoView {
                 PrincipalType::Role => {
                     format!("Update '{name}' Role")
                 }
+                PrincipalType::ApiKey => {
+                    format!("Update '{name}' API Key")
+                }
+                PrincipalType::OauthClient => {
+                    format!("Update '{name}' OAuth Client")
+                }
                 _ => unreachable!(),
             }
         } else {
@@ -320,6 +342,8 @@ pub fn PrincipalEdit() -> impl IntoView {
                 PrincipalType::Tenant => "Create Tenant",
                 PrincipalType::Domain => "Create Domain",
                 PrincipalType::Role => "Create Role",
+                PrincipalType::ApiKey => "Create API Key",
+                PrincipalType::OauthClient => "Create OAuth Client",
                 _ => unreachable!(),
             }
             .to_string()
@@ -362,7 +386,10 @@ pub fn PrincipalEdit() -> impl IntoView {
                                 <Tab tabs=Signal::derive(move || {
                                     vec![
                                         Some("Details".to_string()),
-                                        matches!(typ, PrincipalType::Individual)
+                                        matches!(
+                                            typ,
+                                            PrincipalType::Individual | PrincipalType::ApiKey
+                                        )
                                             .then_some("Authentication".to_string()),
                                         matches!(
                                             typ,
@@ -371,14 +398,19 @@ pub fn PrincipalEdit() -> impl IntoView {
                                             .then_some("Limits".to_string()),
                                         (!matches!(
                                             typ,
-                                            PrincipalType::Tenant | PrincipalType::Domain
+                                            PrincipalType::Tenant
+                                            | PrincipalType::Domain
+                                            | PrincipalType::OauthClient
+                                            | PrincipalType::ApiKey
                                         ))
                                             .then_some("Memberships".to_string()),
                                         matches!(
                                             typ,
                                             PrincipalType::Individual
+                                            | PrincipalType::Group
                                             | PrincipalType::Role
                                             | PrincipalType::Tenant
+                                            | PrincipalType::ApiKey
                                         )
                                             .then_some("Permissions".to_string()),
                                     ]
@@ -390,6 +422,9 @@ pub fn PrincipalEdit() -> impl IntoView {
                                             label=Signal::derive(move || {
                                                 match selected_type.get() {
                                                     PrincipalType::Individual => "Login name",
+                                                    PrincipalType::Domain => "Domain name",
+                                                    PrincipalType::ApiKey => "Key Id",
+                                                    PrincipalType::OauthClient => "Client Id",
                                                     _ => "Name",
                                                 }
                                                     .to_string()
@@ -400,6 +435,7 @@ pub fn PrincipalEdit() -> impl IntoView {
                                                 placeholder=Signal::derive(move || {
                                                     match selected_type.get() {
                                                         PrincipalType::Individual => "Login name",
+                                                        PrincipalType::Domain => "example.org",
                                                         _ => "Short Name",
                                                     }
                                                         .to_string()
@@ -413,7 +449,9 @@ pub fn PrincipalEdit() -> impl IntoView {
                                             stacked=true
                                             label=Signal::derive(move || {
                                                 match selected_type.get() {
-                                                    PrincipalType::Individual => "Name",
+                                                    PrincipalType::Individual | PrincipalType::OauthClient => {
+                                                        "Name"
+                                                    }
                                                     _ => "Description",
                                                 }
                                                     .to_string()
@@ -439,7 +477,10 @@ pub fn PrincipalEdit() -> impl IntoView {
 
                                             hide=Signal::derive(move || {
                                                 is_tenant
-                                                    || matches!(selected_type.get(), PrincipalType::Tenant)
+                                                    || matches!(
+                                                        selected_type.get(),
+                                                        PrincipalType::Tenant | PrincipalType::OauthClient
+                                                    )
                                             })
                                         >
 
@@ -467,6 +508,7 @@ pub fn PrincipalEdit() -> impl IntoView {
                                                     PrincipalType::Individual
                                                     | PrincipalType::Group
                                                     | PrincipalType::List
+                                                    | PrincipalType::OauthClient
                                                 )
                                             })
                                         >
@@ -503,7 +545,9 @@ pub fn PrincipalEdit() -> impl IntoView {
                                             hide=Signal::derive(move || {
                                                 !matches!(
                                                     selected_type.get(),
-                                                    PrincipalType::Tenant | PrincipalType::Domain
+                                                    PrincipalType::Tenant
+                                                    | PrincipalType::Domain
+                                                    | PrincipalType::OauthClient
                                                 )
                                             })
                                         >
@@ -511,6 +555,21 @@ pub fn PrincipalEdit() -> impl IntoView {
                                             <InputText
                                                 element=FormElement::new("picture", data)
                                                 disabled=!is_enterprise
+                                            />
+                                        </FormItem>
+
+                                        <FormItem
+                                            stacked=true
+                                            label="Redirect URIs"
+                                            hide=Signal::derive(move || {
+                                                !matches!(selected_type.get(), PrincipalType::OauthClient)
+                                            })
+                                        >
+
+                                            <StackedInput
+                                                element=FormElement::new("urls", data)
+                                                placeholder="URI"
+                                                add_button_text="Add URI".to_string()
                                             />
                                         </FormItem>
 
@@ -556,6 +615,36 @@ pub fn PrincipalEdit() -> impl IntoView {
                                                 element=FormElement::new("app_passwords", data)
                                                 add_button_text="Add password".to_string()
                                             />
+
+                                        </FormItem>
+
+                                        <FormItem
+                                            stacked=true
+                                            label="Key"
+
+                                            hide=Signal::derive(move || {
+                                                !matches!(selected_type.get(), PrincipalType::ApiKey)
+                                            })
+                                        >
+
+                                            <span class="block font-semibold font-mono text-gray-800 dark:text-gray-200">
+                                                {move || {
+                                                    let data = data.get();
+                                                    let name = data.value::<String>("name").unwrap_or_default();
+                                                    let secret = data
+                                                        .value::<String>("api_secret")
+                                                        .unwrap_or_default();
+                                                    (!name.is_empty() && !secret.is_empty())
+                                                        .then(|| {
+                                                            format!(
+                                                                "api_{}",
+                                                                general_purpose::STANDARD
+                                                                    .encode(format!("{}:{}", name, secret).as_bytes()),
+                                                            )
+                                                        })
+                                                }}
+
+                                            </span>
 
                                         </FormItem>
 
@@ -644,6 +733,16 @@ pub fn PrincipalEdit() -> impl IntoView {
                                         >
 
                                             <InputText element=FormElement::new("max_roles", data)/>
+                                        </FormItem>
+                                        <FormItem
+                                            stacked=true
+                                            label="Maximum number of API Keys"
+                                            hide=Signal::derive(move || {
+                                                !matches!(selected_type.get(), PrincipalType::Tenant)
+                                            })
+                                        >
+
+                                            <InputText element=FormElement::new("max_api_keys", data)/>
                                         </FormItem>
 
                                     </FormSection>
@@ -750,8 +849,10 @@ pub fn PrincipalEdit() -> impl IntoView {
                                                 !matches!(
                                                     selected_type.get(),
                                                     PrincipalType::Individual
+                                                    | PrincipalType::Group
                                                     | PrincipalType::Tenant
                                                     | PrincipalType::Role
+                                                    | PrincipalType::ApiKey
                                                 )
                                             })
                                         >
@@ -776,35 +877,114 @@ pub fn PrincipalEdit() -> impl IntoView {
                                                 !matches!(
                                                     selected_type.get(),
                                                     PrincipalType::Individual
+                                                    | PrincipalType::Group
                                                     | PrincipalType::Role
                                                     | PrincipalType::Tenant
+                                                    | PrincipalType::ApiKey
                                                 )
                                             })
                                         >
 
-                                            <CheckboxGroup
-                                                element=FormElement::new("enabled-permissions", data)
-                                                options=permissions
-                                            />
+                                            <div class="grid space-y-2">
+                                                {move || {
+                                                    let form_data = data.get();
+                                                    let enabled_permissions = form_data
+                                                        .array_value("enabled-permissions")
+                                                        .collect::<AHashSet<_>>();
+                                                    let disabled_permissions = form_data
+                                                        .array_value("disabled-permissions")
+                                                        .collect::<AHashSet<_>>();
+                                                    PERMISSIONS
+                                                        .iter()
+                                                        .map(|(id, name)| {
+                                                            view! {
+                                                                <div class="flex flex-col space-y-3 p-3 w-full bg-white border border-gray-200 rounded-lg focus:border-blue-500 focus:ring-blue-500 dark:bg-neutral-900 dark:border-neutral-700">
+                                                                    <div class="flex justify-between items-center">
+                                                                        <span
+                                                                            id="hs-radio-delete-description"
+                                                                            class="block text-sm text-gray-600 dark:text-neutral-500"
+                                                                        >
+                                                                            {name.to_string()}
+                                                                        </span>
 
-                                        </FormItem>
-                                        <FormItem
-                                            label="Disabled"
-                                            hide=Signal::derive(move || {
-                                                !matches!(
-                                                    selected_type.get(),
-                                                    PrincipalType::Individual
-                                                    | PrincipalType::Role
-                                                    | PrincipalType::Tenant
-                                                )
-                                            })
-                                        >
+                                                                        <div class="flex gap-x-6">
+                                                                            <div class="flex">
+                                                                                <input
+                                                                                    type="radio"
+                                                                                    name=id.to_string()
+                                                                                    class="shrink-0 mt-0.5 border-gray-200 rounded-full text-blue-600 focus:ring-blue-500 disabled:opacity-50 disabled:pointer-events-none dark:bg-neutral-800 dark:border-neutral-700 dark:checked:bg-blue-500 dark:checked:border-blue-500 dark:focus:ring-offset-gray-800"
+                                                                                    id=format!("{id}-on")
+                                                                                    prop:checked=enabled_permissions.contains(id)
+                                                                                    on:input=move |_| {
+                                                                                        data.update_untracked(|data| {
+                                                                                            data.array_push(
+                                                                                                "enabled-permissions",
+                                                                                                id.to_string(),
+                                                                                                true,
+                                                                                            );
+                                                                                            data.array_delete_item("disabled-permissions", id);
+                                                                                        });
+                                                                                    }
+                                                                                />
 
-                                            <CheckboxGroup
-                                                element=FormElement::new("disabled-permissions", data)
-                                                options=permissions
-                                            />
+                                                                                <label class="text-sm text-gray-500 ms-2 dark:text-neutral-400">
+                                                                                    On
+                                                                                </label>
+                                                                            </div>
 
+                                                                            <div class="flex">
+                                                                                <input
+                                                                                    type="radio"
+                                                                                    name=id.to_string()
+                                                                                    class="shrink-0 mt-0.5 border-gray-200 rounded-full text-blue-600 focus:ring-blue-500 disabled:opacity-50 disabled:pointer-events-none dark:bg-neutral-800 dark:border-neutral-700 dark:checked:bg-blue-500 dark:checked:border-blue-500 dark:focus:ring-offset-gray-800"
+                                                                                    id=format!("{id}-off")
+                                                                                    prop:checked=disabled_permissions.contains(id)
+                                                                                    on:input=move |_| {
+                                                                                        data.update_untracked(|data| {
+                                                                                            data.array_push(
+                                                                                                "disabled-permissions",
+                                                                                                id.to_string(),
+                                                                                                true,
+                                                                                            );
+                                                                                            data.array_delete_item("enabled-permissions", id);
+                                                                                        });
+                                                                                    }
+                                                                                />
+
+                                                                                <label class="text-sm text-gray-500 ms-2 dark:text-neutral-400">
+                                                                                    Off
+                                                                                </label>
+                                                                            </div>
+
+                                                                            <div class="flex">
+                                                                                <input
+                                                                                    type="radio"
+                                                                                    name=id.to_string()
+                                                                                    class="shrink-0 mt-0.5 border-gray-200 rounded-full text-blue-600 focus:ring-blue-500 disabled:opacity-50 disabled:pointer-events-none dark:bg-neutral-800 dark:border-neutral-700 dark:checked:bg-blue-500 dark:checked:border-blue-500 dark:focus:ring-offset-gray-800"
+                                                                                    id=format!("{id}-inherit")
+                                                                                    prop:checked=!enabled_permissions.contains(id)
+                                                                                        && !disabled_permissions.contains(id)
+                                                                                    on:input=move |_| {
+                                                                                        data.update_untracked(|data| {
+                                                                                            data.array_delete_item("disabled-permissions", id);
+                                                                                            data.array_delete_item("enabled-permissions", id);
+                                                                                        });
+                                                                                    }
+                                                                                />
+
+                                                                                <label class="text-sm text-gray-500 ms-2 dark:text-neutral-400">
+                                                                                    Default
+                                                                                </label>
+                                                                            </div>
+                                                                        </div>
+                                                                    </div>
+                                                                </div>
+                                                            }
+                                                        })
+                                                        .collect_view()
+                                                }}
+
+                                            </div>
                                         </FormItem>
 
                                     </FormSection>
@@ -879,11 +1059,12 @@ impl FormData {
                             2 => "max_groups",
                             3 => "max_resources",
                             4 => "max_locations",
-                            5 => "max_lists",
-                            6 => "max_other",
-                            7 => "max_domains",
-                            8 => "max_tenants",
-                            9 => "max_roles",
+                            6 => "max_lists",
+                            7 => "max_other",
+                            8 => "max_domains",
+                            9 => "max_tenants",
+                            10 => "max_roles",
+                            11 => "max_api_keys",
                             _ => continue,
                         };
 
@@ -912,6 +1093,7 @@ impl FormData {
                 "disabled-permissions",
                 principal.disabled_permissions.as_string_list(),
             ),
+            ("urls", principal.urls.as_string_list()),
         ] {
             self.array_set(key, list.iter());
         }
@@ -922,6 +1104,8 @@ impl FormData {
                 app_passwords.push(app);
             } else if secret.is_otp_auth() {
                 self.set("otpauth_url", secret);
+            } else if default_type == PrincipalType::ApiKey {
+                self.set("api_secret", secret);
             }
         }
         if !app_passwords.is_empty() {
@@ -937,7 +1121,10 @@ impl FormData {
             }
             if let Some(password) = self.value::<String>("password") {
                 secrets.push(sha512_crypt::hash(password).unwrap());
+            } else if let Some(password) = self.value::<String>("api_secret") {
+                secrets.push(password);
             }
+
             if let Some(otpauth_url) = self.value::<String>("otpauth_url") {
                 secrets.push(otpauth_url);
             }
@@ -968,6 +1155,7 @@ impl FormData {
                 ("lists", &mut principal.lists),
                 ("enabled-permissions", &mut principal.enabled_permissions),
                 ("disabled-permissions", &mut principal.disabled_permissions),
+                ("urls", &mut principal.urls),
             ] {
                 *list = PrincipalValue::StringList(
                     self.array_value(key).map(|m| m.to_string()).collect(),
@@ -989,11 +1177,13 @@ impl FormData {
                     "max_groups",
                     "max_resources",
                     "max_locations",
+                    "max_admins",
                     "max_lists",
                     "max_other",
                     "max_domains",
                     "max_tenants",
                     "max_roles",
+                    "max_api_keys",
                 ]
                 .iter()
                 .map(|f| self.value::<u64>(f).unwrap_or_default())
@@ -1028,6 +1218,10 @@ impl Builder<Schemas, ()> {
                 [Transformer::Trim, Transformer::Lowercase],
                 [Validator::IsEmail],
             )
+            .build()
+            .new_field("urls")
+            .typ(Type::Array)
+            .input_check([Transformer::Trim], [Validator::IsUrl])
             .build()
             .new_field("description")
             .typ(Type::Input)